be/fix: 추천 도서 엔티티 BookId 자료형 맞춰 수정 (#42)

+++ BE/Spring/audisay/src/main/java/com/palja/audisay/domain/recommendation/entity/CategoryBook.java
@@ -16,5 +16,5 @@ public class CategoryBook {
 	@Id
 	private String id;
 	private String groupId;
-	private List<Integer> bookList;
+	private List<Long> bookList;
 }
